Add variables for setting aws access and secret keys
Prior to this commit, I got an error stating that my ova couldn't
be uploaded to S3 during the amazon-import post-processor.

I suspect this is because my credentials set in ~/.aws/credentials
were not looked up as I had expected.

After this commit, I will be setting the keys in a variables.json
file that I can pass to packer build.

Added a .gitignore entry for varibles.json as I really don't want
to accidentally upload my credentials to github.